Use resty.http to replace ngx-socket, to handle chunked HTTP payload.
chunked Data is sent in a series of chunks. The Content-Length header is
omitted in this case and at the beginning of each chunk you need to add the
length of the current chunk in hexadecimal format, followed by '\r\n' and then
the chunk itself, followed by another '\r\n'. The terminating chunk is a
regular chunk, with the exception that its length is zero.

See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Transfer-Encoding